requirements.txt
asgiref==3.5.2
certifi==2022.6.15.1
charset-normalizer==2.1.1
distlib==0.3.6
Django==4.1
django-filter==22.1
django-jazzmin==2.5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
docopt==0.6.2
filelock==3.8.0
idna==3.3
pipreqs==0.4.11
platformdirs==2.5.2
PyJWT==2.4.0
pytz==2018.7
requests==2.28.1
six==1.12.0
sqlparse==0.4.2
urllib3==1.26.12
virtualenv==20.16.3
yarg==0.1.9
